feat: introduced high_level and low_level model APIs
