+++ ramsey/_src/gaussian_process/data.py
@@ -27,11 +27,11 @@ def sample_from_sine(key, n_samples, sigma_noise, frequency = 1, amplitude=1, of
 
 def sample_from_gp_with_rbf_kernel(key, n_samples, sigma_noise, sigma = 1, rho = 1, x_min = -10, x_max = 10):
 
-    print('Sample from GP with RBF Kernel')
-    print('  n_samples = %d' % (n_samples))
-    print('  sigma = %.3f' % (sigma))
-    print('  rho = %.3f' % (rho))
-    print('  noise stddev = %.3f' % (sigma_noise**2))
+    print('  Sample from GP with RBF Kernel')
+    print('    n_samples = %d' % (n_samples))
+    print('    sigma = %.3f' % (sigma))
+    print('    rho = %.3f' % (rho))
+    print('    noise stddev = %.3f' % (sigma_noise**2))
 
 
     x = random.uniform(key, (n_samples,1)) * (x_max - x_min) + x_min
