+++ ramsey/models/high_level.py
@@ -1,0 +1,3 @@
+from ramsey._src.gaussian_process.gaussian_process_wrapper import GP
+
+__all__ = ["GP"]
